% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alcohol_expenditure}
\alias{alcohol_expenditure}
\title{Alcohol Expenditure}
\format{
A data frame with 48 rows and 5 columns:
\describe{
  \item{year}{year}
  \item{tobacco_category}{one of 2 categories of tobacco, cigarettes and hrt}
  \item{exp_alcohol_mn}{total spending on licit tobacco (£millions)}
  \item{exp_alcohol_mn_bp}{total spending on licit tobacco in basic prices (£millions)}
  \item{spend_illicit_mn}{total spending on illicit tobacco (£millions)}
}
}
\source{
\href{https://www.gov.uk/government/statistics/alcohol-bulletin}{Alcohol Bulletin}
}
\usage{
alcohol_expenditure
}
\description{
Data on expenditure on beer, cider, wine, and spirits between 2010 and 2021. Expenditure data are
obtained from the \href{https://www.publichealthscotland.scot/publications/mesas-monitoring-report-2022/}{2022 MESAS monitoring report}.
Data last accessed July 2024. The \href{https://www.gov.uk/government/statistics/alcohol-bulletin}{HMRC Alcohol Bulletin} is used to
estimate total tax receipts from alcohol, used to calculate expenditure measured in basic prices.
}
\keyword{datasets}
