% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdohio.R
\name{cdohio}
\alias{cdohio}
\title{Input-Output Model}
\usage{
cdohio(
  year = 2020,
  year_io = 2020,
  reallocate_prop = 1,
  excluded_products = NULL,
  change_food = rep(0, 19),
  change_gambling = rep(0, 9),
  change_tobacco_licit = rep(0, 2),
  change_tobacco_illicit = rep(0, 2),
  change_alcohol = rep(0, 4)
)
}
\arguments{
\item{year}{Numeric. Year of expenditure, tax, and labour market outcomes data to use.}

\item{year_io}{Numeric. Year of input-output tables to use (select one from 2017. 2018, 2019, or 2020) - default is 2020.}

\item{reallocate_prop}{Numeric. Proportion of total change in spending reallocated to other products (0 to 1) - default is 1.}

\item{excluded_products}{Character vector. Products to exclude from reallocation. The products excluded can be any from
c("alcohol","tobacco","food","gambling") - default is to exclude all four categories.}

\item{change_food}{Numeric integer. Percentage change in expenditure on 19 categories of food;
1. Milk, 2. Cheese, 3. Carcase meat, 4. Non-carcase meat and meat products, 5. Fish, 6. Eggs,
7. Fats, 8. Sugar and preserves, 9. Potatoes, 10. Fruit and veg, 11. Bread, 12. Flour, 13. Cakes, buns, and pastries,
14. Biscuits and crispbreads, 15. Cereals and cereal products, 16. Beverages, 17. Other food and drink, 18. Soft drinks,
19. Confectionary.}

\item{change_gambling}{Numeric integer. Percentage change in expenditure (gross gambling yield) on 9
gambling categories (arcades, betting (non-remote), bingo (non-remote), casino (non-remote),
betting (remote), bingo (remote), casino (remote), lotteries, the National Lottery)}

\item{change_tobacco_licit}{Numeric vector. Percentage change in expenditure on 2
licit tobacco product categories (cigarettes, handrolled tobacco)}

\item{change_tobacco_illicit}{Numeric vector. Percentage change in expenditure on 2
illicit tobacco product categories (cigarettes, handrolled tobacco)}

\item{change_alcohol}{Numeric vector. Percentage change in expenditure on 4 alcohol
product categories (beer, cider, spirits, wine)}
}
\value{
List object
}
\description{
A wrapper function to produce input-output analysis of the effects on the UK
economy of changes in the demand
}
\examples{

\dontrun{

}
}
