% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coicop_categories}
\alias{coicop_categories}
\title{Classification of Individual Consumption According to Purpose}
\format{
A data frame with 36 rows and 3 columns:
\describe{
  \item{index}{Position of the COICOP category in the vector}
  \item{code}{COICOP household expenditure category code}
  \item{label}{Description}
}
}
\usage{
coicop_categories
}
\description{
The Classification of Individual Consumption According to Purpose (COICOP) is a 36-category
classification of household expenditures. It is mapped onto the 105 CPA product categories in the
supply and use table and is used as the basis for reallocation of spending in the model.
}
\keyword{datasets}
