% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpa_categories}
\alias{cpa_categories}
\title{CPA Product Categories}
\format{
A data frame with 105 rows and 2 columns:
\describe{
  \item{CPA}{CPA product code}
  \item{Product}{Product description}
}
}
\usage{
cpa_categories
}
\description{
A lookup table of the 105 product categories used in the input-output analytical tables
}
\keyword{datasets}
