% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpa_conversion_matrix}
\alias{cpa_conversion_matrix}
\title{CPA Conversion Matrix}
\format{
A data frame with 48 rows and 5 columns:
\describe{
  \item{CPA_hhfce}{CPA codes in the }
  \item{tobacco_category}{one of 2 categories of tobacco, cigarettes and hrt}
  \item{exp_alcohol_mn}{total spending on licit tobacco (£millions)}
  \item{exp_alcohol_mn_bp}{total spending on licit tobacco in basic prices (£millions)}
  \item{spend_illicit_mn}{total spending on illicit tobacco (£millions)}
}
}
\usage{
cpa_conversion_matrix
}
\description{
A matrix to map the 103 CPA categories in the HHFCE component of the ONS supply and use tables to the 105 CPA categories
in the supply table and input-output analytical tables. The difference in dimensions is due to the HHFCE tables (i) merging
CPA_27 and CPA_28 into a combined electrical equipment plus machinery and equipment n.e.c category, and (ii) merging CPA H53
and CPA I55 into a combined postal and courier services plus accommodation services. The mapping splits expenditure on both
of these combined categories equally between the two component products.
}
\keyword{datasets}
