% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpa_to_sector}
\alias{cpa_to_sector}
\title{CPA Conversion Matrix}
\format{
A data frame with 48 rows and 4 columns:
\describe{
  \item{CPA}{CPA codes in the input-output tables (105 categories)}
  \item{Product}{CPA product description}
  \item{sector}{20-category classification of the 105 CPA categories into SIC sectors}
  \item{sector_broad}{12-category version of the classification}
}
}
\usage{
cpa_to_sector
}
\description{
A mapping of the 105 CPA product categories onto broad sectors based on the Standard Industrial Classification
(SIC-2007).
}
\keyword{datasets}
