% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EconImpactCalc.R
\name{EconImpactCalc}
\alias{EconImpactCalc}
\title{Calculate Economic Impacts}
\usage{
EconImpactCalc(
  year = 2019,
  year_io = 2019,
  base = 2019,
  input_vector,
  alcohol_off_tax = 0,
  alcohol_on_tax = 0,
  tobacco_tax = 0
)
}
\arguments{
\item{year}{Numeric. Year of expenditure, tax, and labour market outcomes data to use - default is 2019.}

\item{year_io}{Numeric. Year of input-output tables to use (select one from 2017. 2018, 2019, or 2020) - default is 2019.}

\item{base}{Numeric. Base year to use for inflation adjustment - default is 2019.}

\item{input_vector}{Numeric vector. The vector of length 105 of changes in final demand in basic prices.}

\item{alcohol_off_tax}{Numeric. Change in tax less subsidies on products for off-trade alcohol calculated from `GenExpenditure()` function.}

\item{alcohol_on_tax}{Numeric. Change in tax less subsidies on products for on-trade alcohol calculated from `GenExpenditure()` function.}

\item{tobacco_tax}{Numeric. Change in tax less subsidies on products for tobacco calculated from `GenExpenditure()` function.}
}
\value{
A list of outputs
}
\description{
Apply the input-output model parameters to a pre-specified vector of changes in final demand for 105 products. Calculate
impacts on output, gross value added, tax receipts to government, employment, and earnings. Three types of effect are
reported - the direct effects, direct + indirect effects, and direct + indirect + induced effects.
}
\examples{

\dontrun{

}
}
