% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{food_expenditure}
\alias{food_expenditure}
\title{Food Expenditure Data}
\format{
A data frame with 152 rows and 4 columns:
\describe{
  \item{year}{year}
  \item{food_category}{one of 19 categories of food consumption}
  \item{cpa}{CPA product category in the input-output model}
  \item{exp_per_year_mn}{annual total expenditure (£millions)}
}
}
\source{
\href{https://www.gov.uk/government/statistical-data-sets/family-food-datasets}{ONS Family Food Datasets}
}
\usage{
food_expenditure
}
\description{
Data on expenditure on broad categories of food between 2015 and 2022.
Data last accessed July 2024.
}
\keyword{datasets}
