% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fte_coefficients}
\alias{fte_coefficients}
\title{Full-Time Equivalent Employment Coefficients}
\format{
An object of class \code{numeric} of length 105.
}
\usage{
fte_coefficients
}
\description{
Coefficients describing the change in employment for a £1mn change in output for each
sector corresponding to the 105 CPA products.
}
\details{
The coefficients are obtained by matching the 105 CPA product categories to the equivalent 105
industries, based on the Standard Industrial Classification (SIC-2007), that the industry-by-industry
input-output tables are based on. The proportion of output in each industry that is composed of
compensation of employees is calculated for each industry from the 2020 industry-by-industry input-output table.
}
\keyword{datasets}
