% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gambling_expenditure}
\alias{gambling_expenditure}
\title{Gambling Expenditure Data}
\format{
A data frame with 135 rows and 3 columns:
\describe{
  \item{year}{year}
  \item{gambling_category}{one of 9 categories of gambling}
  \item{exp_per_year_mn}{total gross gambling yield (GGY) (£millions)}
}
}
\source{
\href{https://www.data.gov.uk/dataset/e7032815-5990-4439-b5c8-8553cf5b7fdd/gb-gambling-industry-statistics-february-2024}{Gambling Commission statistics}
}
\usage{
gambling_expenditure
}
\description{
Data on the gross gambling yield (GGY) for 9 categories of gambling spending between
2008 and 2022. Data last accessed July 2024.
}
\keyword{datasets}
