% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{govt_spending}
\alias{govt_spending}
\title{Government spending distribution}
\format{
A data frame with 105 rows and 6 columns:
\describe{
  \item{CPA}{CPA codes in the input-output tables (105 categories)}
  \item{Product}{Product description}
  \item{govt_2017}{proportion of central government spending by CPA category (2017)}
  \item{govt_2018}{proportion of central government spending by CPA category (2018)}
  \item{govt_2019}{proportion of central government spending by CPA category (2019)}
  \item{govt_2020}{proportion of central government spending by CPA category (2020)}
}
}
\usage{
govt_spending
}
\description{
Vectors of central government spending distributions across the 105 CPA categories in the input-output tables
}
\keyword{datasets}
