% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GovtSpendVec.R
\name{GovtSpendVector}
\alias{GovtSpendVector}
\title{Generate Government Spending Change Input Vector}
\usage{
GovtSpendVector(year_io = 2019, govt_spend = 100, govt_allocation = 1)
}
\arguments{
\item{year_io}{Numeric. Year of input-output tables to use (select one from 2017. 2018, 2019, or 2020) - default is 2019.}

\item{govt_spend}{Numeric. Changes in central government spending.}

\item{govt_allocation}{Numeric. options for the allocation of changes in central government spending. 1 for pro-rata
allocation according to the input-output tables. 2 for full allocation to health (CPA_Q86).}
}
\value{
A vector of length 105
}
\description{
Construct a vector of changes in government spending to add as inputs to
changes in household final demand and include in the input-output model.
}
\examples{

\dontrun{

}
}
