% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{income_tax_params}
\alias{income_tax_params}
\title{Income Tax Parameters 2010-2022}
\format{
A data table with 13 rows and 13 columns:
}
\source{
\href{https://www.gov.uk/government/collections/tax-structure-and-parameters-statistics}{Tax structure and parameters statistics}
}
\usage{
income_tax_params
}
\description{
Data table of income tax and national insurance parameters - thresholds and tax rates - used for calculating tax paid on employment incomes
based on annual earnings. Data provided for 2010 - 2022
}
\keyword{datasets}
