% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inputoutput_2022}
\alias{inputoutput_2022}
\title{UK Input-Output Table (2022)}
\format{
A list object with 12 items:
\describe{
  \item{product_categories_io}{105 CPA product categories in the input-output table}
  \item{product_categories_hhfce}{103 CPA product categories in the HHFCE table}
  \item{supply}{the supply table}
  \item{iot}{the input-output flowtable showing inter-industry use of products}
  \item{A}{the matrix of technical coefficients - the input-output table elements as a proportion of total output}
  \item{L}{the Leontief inverse (type 1)}
  \item{L2}{the Leontief inverse (type 2)}
  \item{output}{total product output in basic prices (£mn)}
  \item{gva}{gross value added (£mn)}
  \item{tax_on_products}{taxes less subsidies on products (£mn)}
  \item{tax_on_production}{taxes less subsidies on production (£mn)}
  \item{gva_coefficients}{Gross value added (GVA) as a proportion of total product output}
  \item{tax_coefficients}{Net tax on production as a proportion of total product output}
  \item{coe_coefficients}{Compensation of employees as a proportion of total product output}
  \item{hhfce_categories}{36 COICOP categories of consumption}
  \item{hhfce}{Correspondence table mapping expenditure in the 36 COICOP categories to the CPA product categories}
  \item{hhfce_totals}{Total expenditure in the 36 COICOP categories of consumption}
}
}
\source{
\href{https://www.ons.gov.uk/economy/nationalaccounts/supplyandusetables/datasets/inputoutputsupplyandusetables}{Supply and use tables}

\href{https://www.ons.gov.uk/economy/nationalaccounts/supplyandusetables/datasets/ukinputoutputanalyticaltablesdetailed}{Input-output tables (product by product)}
}
\usage{
inputoutput_2022
}
\description{
Product by product Input-Output Analytical Tables with 105 product categories for 2022
from the Office for National Statistics, with supporting data from the \href{https://www.ons.gov.uk/economy/nationalaccounts/supplyandusetables/datasets/inputoutputsupplyandusetables}{supply and use tables}.
including the supply table and household final consumption expenditure (HHFCE).
}
\keyword{datasets}
