% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lfs_empl_data}
\alias{lfs_empl_data}
\title{Annual Employment}
\format{
A data table with 105 rows and 14 columns:
\describe{
  \item{SIC}{105 Standard Industrial Classification for 2007 (SIC-2007) codes corresponding to 105 CPA categories}
  \item{SIC_Industry}{Industry descriptions}
  \item{empl_2017}{total employment for 2017}
  \item{fte_2017}{full-time equivalent employment for 2017}
  \item{empl_2018}{total employment for 2018}
  \item{fte_2018}{full-time equivalent employment for 2018}
  \item{empl_2019}{total employment for 2019}
  \item{fte_2019}{full-time equivalent employment for 2019}
  \item{empl_2020}{total employment for 2020}
  \item{fte_2020}{full-time equivalent employment for 2020}
  \item{empl_2021}{total employment for 2021}
  \item{fte_2021}{full-time equivalent employment for 2021}
  \item{empl_2022}{total employment for 2022}
  \item{fte_2022}{full-time equivalent employment for 2022}
}
}
\source{
\href{https://beta.ukdataservice.ac.uk/datacatalogue/series/series?id=2000026}{Labour Force Survey. DOI: 10.5255/UKDA-Series-2000026}
}
\usage{
lfs_empl_data
}
\description{
Employment by sector corresponding to the 105 products in the model. Employment data are available for
years 2017 to 2022 as both total employment and full-time-equivalent (FTE) employment. FTE employment is
constructed as a weighted sum of employment where full-time workers have a weight of 1 and part-time workers
a weight of 0.5. Usual weekly hours worked of less than 30 hours is considered part-time.
}
\keyword{datasets}
