% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lfs_wage_data}
\alias{lfs_wage_data}
\title{Average Annual Earnings}
\format{
A data table with 105 rows and 8 columns:
\describe{
  \item{SIC}{Standard Industrial Classification for 2007 (SIC-2007) codes}
  \item{SIC_Industry}{103 CPA product categories in the HHFCE table}
  \item{earn_2017}{gross average annual earnings for 2017}
  \item{earn_2018}{gross average annual earnings for 2018}
  \item{earn_2019}{gross average annual earnings for 2019}
  \item{earn_2020}{gross average annual earnings for 2020}
  \item{earn_2021}{gross average annual earnings for 2021}
  \item{earn_2022}{gross average annual earnings for 2022}
}
}
\usage{
lfs_wage_data
}
\description{
Gross average annual earnings by sector associated with each product. Average earnings data are available for
years 2017 to 2022. Data are sourced from the Labour Force Survey (LFS), and averages for each year are calculated from
a rolling three-year period to ensure sufficient sample size to calculate an average for each sector, e.g.
earnings for 2017 are calculated using LFS data from 2016-2018.
}
\keyword{datasets}
