% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multipliers.R
\name{Multipliers}
\alias{Multipliers}
\title{Multipliers}
\usage{
Multipliers(year_io = 2020)
}
\arguments{
\item{year_io}{Numeric. Year of input-output tables to use (select one from 2017. 2018, 2019, or 2020) - default is 2020.}
}
\value{
A data table of multipliers
}
\description{
Calculate multipliers for a selected year of input-output product by product tables to show the change in output and
gross value added (GVA) for a one-unit change in final demand.
}
\examples{

\dontrun{

}
}
