% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rpi}
\alias{rpi}
\title{RPI Inflation}
\format{
A data table with 37 rows and 2 columns:
\describe{
  \item{year}{year}
  \item{rpi_index}{inflation index (base = July 1987)}
}
}
\source{
\href{https://www.ons.gov.uk/economy/inflationandpriceindices/timeseries/chaw/mm23}{ONS inflation and price indices}
}
\usage{
rpi
}
\description{
Annual inflation from the Retail Price Index (RPI) for years 1987 - 2023
}
\keyword{datasets}
