% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxCalc.R
\name{TaxCalc}
\alias{TaxCalc}
\title{Tax Calculator}
\usage{
TaxCalc(
  earn_data = cdohio.mod::lfs_wage_data,
  tax_data = cdohio.mod::income_tax_params,
  year = 2019
)
}
\arguments{
\item{earn_data}{Data table. Package data containing average earnings by 105 CPA sectors.}

\item{tax_data}{Data table. Package data containing the necessary parameters to calculate
income tax and national insurance contributions from annual earnings data}

\item{year}{Integer. Select the year (from 2016 to 2022) of earnings and employment data to use in the analysis - default is 2019.}
}
\value{
Data table of earnings and tax impact estimates
}
\description{
Simulate an earnings distribution for each sector in the model and calculate income
tax and national insurance contributions (NICs) estimates
}
\examples{

\dontrun{


}
}
