% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tobacco_expenditure}
\alias{tobacco_expenditure}
\title{Tobacco Expenditure Data}
\format{
A data frame with 8 rows and 5 columns:
\describe{
  \item{year}{year}
  \item{tobacco_category}{one of 2 categories of tobacco, cigarettes and hrt}
  \item{spend_licit_mn}{total spending on licit tobacco (£millions)}
  \item{spend_illicit_mn}{total spending on illicit tobacco (£millions)}
  \item{excise_pct}{excise tax as a percentage of total expenditure}
  \item{tax_pct}{excise tax and VAT as a percentage of total expenditure}
}
}
\source{
\href{https://www.gov.uk/government/statistics/tobacco-bulletin}{Tobacco Bulletin}
}
\usage{
tobacco_expenditure
}
\description{
Data on expenditure for 2 categories of tobacco spending (cigarettes and handrolled tobacco) between
2019 and 2022. Data last accessed July 2024.
}
\details{
Total licit spending data are constructed by calculating the percentage of
the retail price for tobacco products which is made up of duty. These figures are then used to adjust
total duty receipts data from the \href{https://www.gov.uk/government/statistics/tobacco-bulletin}{HMRC Tobacco Bulletin} to
a figure for total expenditure. Data on prices were obtained from \href{https://www.ons.gov.uk/economy/inflationandpriceindices/datasets/consumerpriceindicescpiandretailpricesindexrpiitemindicesandpricequotes}{ONS consumer price inflation price quotes}
data for HRT, and from the Sheffield Tobacco and Alcohol Policy Model (STAPM) for cigarettes in 2017 adjusted for other years using the
\href{https://www.ons.gov.uk/economy/inflationandpriceindices/timeseries/l7aq/mm23}{CPI index for cigarettes}.

Illicit spending is calculated using estimates of the total volume of illicit tobacco published by HMRC in the \href{https://www.gov.uk/government/statistics/measuring-tax-gaps-tables}{measuring tax gaps tables}.
Consumption is multiplied by the price of illicit tobacco, which is assumed to be half of the price of licit tobacco.
}
\keyword{datasets}
