% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alcohol_expenditure}
\alias{alcohol_expenditure}
\title{Alcohol Expenditure}
\format{
A data frame with 48 rows and 4 columns:
\describe{
  \item{year}{year}
  \item{alcohol_category}{one of 4 categories of alcohol; beer, cider, wine, and spirits}
  \item{exp_alcohol_mn}{total spending on alcohol (£millions)}
  \item{tax_pct}{excise tax and VAT as a percentage of total expenditure}
}
}
\source{
\href{https://www.gov.uk/government/statistics/alcohol-bulletin}{Alcohol Bulletin}
}
\usage{
alcohol_expenditure
}
\description{
Data on expenditure on beer, cider, wine, and spirits between 2010 and 2021. Expenditure data are
obtained from the \href{https://www.publichealthscotland.scot/publications/mesas-monitoring-report-2022/}{2022 MESAS monitoring report}.
Data last accessed July 2024. The \href{https://www.gov.uk/government/statistics/alcohol-bulletin}{HMRC Alcohol Bulletin} is used to
estimate total tax receipts from alcohol, used to calculate expenditure measured in basic prices.
}
\keyword{datasets}
