% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EconImpactCalc.R
\name{EconImpactCalc}
\alias{EconImpactCalc}
\title{Calculate Economic Impacts}
\usage{
EconImpactCalc(year = 2020, input_vector)
}
\arguments{
\item{year}{Numeric. Year of input-output tables to use (select one from 2017. 2018, 2019, or 2020) - default is 2020.}

\item{input_vector}{Numeric vector. The vector of length 105 of changes in final demand in basic prices}
}
\value{
A list of outputs
}
\description{
Apply the input-output model parameters to a pre-specified vector of changes in final demand for 105 products. Calculate
impacts on output, gross value added, tax receipts to government, employment, and earnings. Three types of effect are
reported - the direct effects, direct + indirect effects, and direct + indirect + induced effects.
}
\examples{

\dontrun{

}
}
