% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{food_cat_lookup}
\alias{food_cat_lookup}
\title{Food Categories}
\format{
A data frame with 19 rows and 2 columns:
\describe{
  \item{food_category}{one of 19 categories of food consumption}
  \item{cpa}{CPA product category in the input-output model}
}
}
\source{
\href{https://www.gov.uk/government/statistical-data-sets/family-food-datasets}{ONS Family Food Datasets}
}
\usage{
food_cat_lookup
}
\description{
Lookup table showing the correspondence between the 19 Family Food categories of Food and the
7 categories in the input-output analytical tables.
}
\keyword{datasets}
