% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputExpenditure.R
\name{InputExpenditure}
\alias{InputExpenditure}
\title{Input Demand Changes}
\usage{
InputExpenditure(input_data, model_type)
}
\arguments{
\item{input_data}{Data table. If model_type = "healthei", it must contain the following columns - `CPA`, `Product`, and `exp_change`}

\item{model_type}{Character. A public health model being used to supply inputs. One of c("none","healthei","stapm")}
}
\value{
A list object
}
\description{
Directly input changes in demand for alcohol, tobacco, gambling. This function is used to directly
input changes in the demand for the unhealthy commodities for which the model calculates economic
impact analyses. Use the `GenExpenditure` function to calculate inputs as proportionate changes using
the expenditure data saved to the model.
}
\details{
Note that for alcohol and tobacco this function also calculates changes in demand in basic prices as well as
in purchaser prices. As alcohol and tobacco are combined in the model, and have very different levels of taxation
relative to retail price, the conversion from purchaser prices to basic prices is done here separately for
alcohol and tobacco rather than within the model, as is the case for other products. Because of this, this function also
calculates the change in tax less subsidies on products for alcohol and tobacco so that this is
accounted for in the model.
}
\examples{

\dontrun{

}
}
