% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lfs_wage_data}
\alias{lfs_wage_data}
\title{Average Annual Earnings}
\format{
A data table with 105 rows and 8 columns:
\describe{
  \item{SIC}{105 Standard Industrial Classification for 2007 (SIC-2007) codes corresponding to 105 CPA categories}
  \item{SIC_Industry}{Industry descriptions}
  \item{earn_2017}{gross average annual earnings for 2017}
  \item{earn_2018}{gross average annual earnings for 2018}
  \item{earn_2019}{gross average annual earnings for 2019}
  \item{earn_2020}{gross average annual earnings for 2020}
  \item{earn_2021}{gross average annual earnings for 2021}
  \item{earn_2022}{gross average annual earnings for 2022}
}
}
\source{
\href{https://beta.ukdataservice.ac.uk/datacatalogue/series/series?id=2000026}{Labour Force Survey. DOI: 10.5255/UKDA-Series-2000026}
}
\usage{
lfs_wage_data
}
\description{
Gross average annual earnings by sector associated with each product. Average earnings data are available for
years 2017 to 2022. Data are sourced from the Labour Force Survey (LFS), and averages for each year are calculated from
a rolling three-year period to ensure sufficient sample size to calculate an average for each sector, e.g.
}
\keyword{datasets}
